[infra] print full exceptions with stacktraces (#5)
Triton compiler errors are currently very hard to read as only the source location is printed. This is due to the actual error message being part of a chained exception in the `__cause__` field, which is not printed by the standard string conversion.

This PR changes the print behavior to use Pythons standard exception printing mechanism that provides both a traceback and the message. This could technically also be implemented for just the triton framework, but felt like an overall improvement