add proper autotuning to kernels  (#59)
In some kernels, I realized that we don't do the "itertools" thing. I
added it to all the currently merged kernels that did not have it. I
managed really big speed ups on many of the kernels.

gemm --> 18 ms to 10 ms
doitgen --> 13 ms to 3 ms
trmm --> 23 ms to 6 ms
go_fast --> 5 ms to 2 ms