[npbench] Fix validation logic for cases where framework does not return enough outputs (#63)
In NPBench, after running a benchmark, a list of outputs is constructed.
This contains the return values of the benchmark implementation as well
as any input arguments listed in the `output_args` configuration key of
the benchmark.
However, the validation function uses `zip` to iterate over pairs of
outputs returned by NumPy (reference) and the framework implementation.

In a case where the arguments to `zip` don't have the same length, the
framework will not validate the remaining elements of the larger list.

In particular, if a kernel has no `output_args`, this makes a kernel
returning `None` pass (example for `durbin`):

```python
def kernel(r):
    pass
```

With the improved validation logic, the kernels `go_fast` and `durbin`
sadly don't pass anymore.