Fix float32 constants (#68)
In NPBench, some input arguments are generated by `initialize`
functions, but some are taken from the JSON file that exists for each
benchmark. The `initialize` function knows about the datatype and can
return scalars specific for that datatype (e.g. a scalar with type
`np.float32` if the datatype is `float32`). It also knows the default
datatype, if none is specified.

However, float constants from the input file are still stored as Python
floats. As it turns out, DaCe is unable to handle these floats when
using datatype `float32` and crashes.
Ideally, we would store the default datatype of each benchmark in its
JSON file. This would allow us to cast the JSON constants to the right
datatype. However, this is a lot of work. Instead, this PR introduces a
hacky solution to detect the datatype that is actually used, and convert
all Python `float`s to that. This fixes some DaCe kernels, like
`channel_flow`, `cavity_flow` and `adi`.